chore: move walk action to walk.rs
