+++ crates/world_state/src/behavior/walk.rs
@@ -174,3 +174,31 @@ pub fn walk_alternatives_subtree() -> Node<Blackboard> {
         action!(stand)
     )
 }
+
+pub fn walk_to_block_position(blackboard: &mut Blackboard) -> Status {
+    if let (Some(block_position), Some(ball), Some(ground_to_field)) = (
+        &blackboard.walk_position,
+        &blackboard.last_ball,
+        blackboard.world_state.robot.ground_to_field,
+    ) {
+        let ball_position = ground_to_field.inverse() * ball.position;
+        let orientation = Orientation2::from_vector(ball_position - *block_position);
+
+        walk_to(
+            blackboard,
+            Pose2::from_parts(*block_position, orientation),
+            blackboard.parameters.walk_speed.blocking,
+            OrientationMode::LookAt {
+                target: ball_position,
+                tolerance: blackboard.parameters.walk_and_stand.orientation_tolerance,
+            },
+            blackboard
+                .parameters
+                .walk_and_stand
+                .normal_distance_to_be_aligned,
+            blackboard.parameters.walk_and_stand.hysteresis,
+        )
+    } else {
+        Status::Failure
+    }
+}
